requirements/base.txt
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-mptt==0.16.0
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
Markdown==3.7
psycopg2-binary==2.9.10
python-decouple==3.8
